requirements.txt
google-generativeai>=0.8.0
python-dotenv>=1.0.0
tqdm>=4.66.0